Data Manager Implemented
Reconfigured the program to utilize a central data management. When the program spawns asteroids, a new data structure is configured for the asteroids, orbs, and cells. When the player selects the asteroid they wish to mine from, its data is stored within the DataManager for future reference from outside scripts.